% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_process.R
\name{ad.correct}
\alias{ad.correct}
\title{Correct allele depth values}
\usage{
ad.correct(het.table, gt.table = NULL, odd.correct = TRUE, verbose = TRUE)
}
\arguments{
\item{het.table}{allele depth table generated from the function hetTgen}

\item{gt.table}{genotype table generated from the function hetTgen}

\item{odd.correct}{logical, to correct for odd number anomalies in AD values. default TRUE}

\item{verbose}{logical. show progress. Default = TRUE}
}
\value{
returns the coverage corrected allele depth table similar to the output of hetTgen
}
\description{
A function to correct depth values where odd number of coverage values due to sequencing anomalies or miss classification where genotype is homozygous and depth values indicate heterozygosity.
The function adds a value of 1 to the allele with the lowest depth value for when odd number anomalies or make the depth value 0 for when miss-classified. The genotype table must be provided for the latter.
}
\examples{
adc<-ad.correct(ADtable)

}
\author{
Piyal Karunarathne
}
