% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADnorm}
\alias{ADnorm}
\title{Normalized allele depth example data}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(ADnorm)
}
\description{
Normalized example SNPs data of Picea abies from Chen et al. 2019
The data has been normalized with TMM
}
\keyword{datasets}
