% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADtable}
\alias{ADtable}
\title{Allele Depth (AD) example data}
\format{
An object of class \code{data.frame} with 3000 rows and 109 columns.
}
\usage{
data(ADtable)
}
\description{
Example SNPs data of Chinook Salmon from Larson et al. et al. 2014.
The data contains only a partial snps data set of RadSeq data after filtering.
}
\references{
\itemize{
\item{"1."}{Larson, W. A., Seeb, L. W., Everett, M. V., Waples, R. K., Templin, W. D., & Seeb, J. E. (2014). Genotyping by sequencing resolves shallow population structure to inform conservation of Chinook salmon (Oncorhynchus tshawytscha). Evolutionary Applications, 7(3), 355-369.}
\item{"2."}{McKinney, G. J., Waples, R. K., Seeb, L. W., & Seeb, J. E. (2017). Paralogs are revealed by proportion of heterozygotes and deviations in read ratios in genotyping‐by‐sequencing data from natural populations. Molecular Ecology Resources, 17(4), 656-669.}
}
}
\keyword{datasets}
