% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.det.R
\name{allele.info}
\alias{allele.info}
\title{Get allele info. for duplicate detection}
\usage{
allele.info(
  X,
  x.norm = NULL,
  method = c("TMM", "TMMex"),
  logratioTrim = 0.3,
  sumTrim = 0.05,
  Weighting = TRUE,
  Acutoff = -1e+10,
  plot.allele.cov = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{X}{a data frame of allele coverage (non-normalized). output of hetTgen()}

\item{x.norm}{a data frame of normalized allele coverage. output of cpm.normal(). If not provided, calculated using X.}

\item{method}{character. method to be used for normalization (see cpm.normal detials). Default="TMM" for data <100,000 snps and "TMMex" for >100,000}

\item{logratioTrim}{numeric. percentage value (0 - 1) of variation to be trimmed in log transformation}

\item{sumTrim}{numeric. amount of trim to use on the combined absolute levels ("A" values) for method="TMM"}

\item{Weighting}{logical, whether to compute (asymptotic binomial precision) weights}

\item{Acutoff}{numeric, cutoff on "A" values to use before trimming}

\item{plot.allele.cov}{logical, plot comparative plots of allele depth coverage in homozygotes and heterozygotes}

\item{verbose}{logical, whether to print progress}

\item{...}{further arguments to be passed to plot}
}
\value{
Returns a data frame of median allele ratio, proportion of heterozygotes, number of heterozygotes, and allele probability at different assumptions with their chi-square significance
}
\description{
The function to calculate allele median ratios, proportion of heterozygotes and allele probability values under different assumptions (see details), and their chi-square significance values for duplicate detection
}
\details{
print("to be added")
}
\examples{
data(ADtable)
AI<-allele.info(ADtable)

}
\author{
Piyal Karunarathne, Pascal Milesi, Qiujie Zhou
}
