% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alleleINF}
\alias{alleleINF}
\title{Allele info example data}
\format{
An object of class \code{data.frame} with 5388 rows and 19 columns.
}
\source{
\href{url}{dataset name}
}
\usage{
data(alleleINF)
}
\description{
Semi-randomly generated data from the function dup.snp.info
Data contains depth and proportion values of 19800 snps
}
\examples{
data(alleleINF)
with(alleleINF,plot(medRatio~propHet))
}
\references{
Karunarathne et al. (2021)
}
\keyword{datasets}
