% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{dup.detect}
\alias{dup.detect}
\title{Detect and plot duplicated snps}
\usage{
dup.detect(d.info, stringency = c("95", "99", "max"), plot = TRUE)
}
\arguments{
\item{d.info}{a data.frame of duplication info generated by dup.snp.info function}

\item{stringency}{character. Confidence interval to be used in filtering duplicates}

\item{plot}{logical. If proportion of heterozygotes vs. allele median ratio should be plotted to visualize the detected duplicates}
}
\value{
a data frame of snps categorized into duplicates and singletons along with allele median ratio and heterozygote and homozygote proportions.
}
\description{
This function uses excess of heterozygotes under HWE depth of coverage vs. no. of heterozygotes dependent expected allele ratios to determine the duplicated snps from singletons.
}
\examples{
data(hets)
d.info <- dup.snp.info(het.table=hets,normalize=FALSE)
duplicated<-dup.detect(d.info,stringency="max",plot=TRUE)

}
\author{
Piyal Karunarathne
}
