% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dup.info}
\alias{dup.info}
\title{dup.info example data}
\format{
An object of class \code{function} of length 1.
}
\source{
\href{url}{dataset name}
}
\usage{
data(dup.info)
}
\description{
Semi-randomly generated data from the function dup.snp.info
Data contains depth and proportion values of 19800 snps
}
\examples{
data(dup.info)
props<-dup.info[,c("PropHomRare","PropHet")]
\donttest{plot(props, pch=19,cex=0.2)}
}
\references{
Karunarathne et al. (2021)
}
\keyword{datasets}
