% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_detect.R
\name{dup.validate}
\alias{dup.validate}
\title{Validate detected deviants/cnvs}
\usage{
dup.validate(d.detect, window.size = 100, scaf.size = 10000)
}
\arguments{
\item{d.detect}{a data frame of detected duplicates or deviants from the outputs of \code{dupGet} or \code{cnv}}

\item{window.size}{numerical. a single value of the desired moving window
size (default \code{100} bp)}

\item{scaf.size}{numerical. scaffold size to be checked. i.e. the chromosome/scaffolds will be split into equal pieces of this size
default=10000}
}
\value{
A data frame of deviant/cnv ratios (column \code{cnv.ratio}) for a split of the chromosome/scaffold given by the \code{scaf.size}; this ratio is an average value of the percentage of deviants/cnvs present within the given \code{window.size} for each split (\code{chromosome/scaffold length/sacf.size}); the start and the end positions of each split is given in the \code{start} and \code{end} columns
}
\description{
This function will validate the detected duplicated-SNPs (deviants/cnvs) using a moving
window approach (see details)
}
\details{
Loci/SNP positions correctly ordered according to a reference
sequence is necessary for this function to work properly. The list of deviants/cnvs provided in the \code{d.detect} will be split into pices of \code{scaf.size} and the number of deviants/cnvs will be counted along each split with a moving window of \code{window.size}. The resulting percentages of deviants/cnvs will be averaged for each scaf.size split; this is the \code{cnv.ratio} column in the output. Thus, ideally, the \code{cnv.ratio} is a measure of how confident the detected deviants/cnvs are in an actual putative duplicated region withing the given \code{scaf.size}. This ratio is sensitive to the picked window size and the scaf.size; as a rule of thumb, it is always good to use a known gene length as the scaf.size, if you need to check a specific gene for the validity of the detected duplicates.
Please also note that this function is still in its \code{beta-testing} phase and also under development for non-mapped reference sequences. Therefore, your feedback and suggestions will be highly appreciated.
}
\author{
Piyal Karunarathne
}
