% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{dupGet}
\alias{dupGet}
\title{Detect duplicates from SNPs}
\usage{
dupGet(
  data,
  test = c("z.het", "z.05", "z.all", "chi.het", "chi.05", "chi.all"),
  intersection = FALSE,
  method = c("fisher", "chi.sq"),
  plot = TRUE,
  verbose = TRUE,
  ...
)
}
\description{
Detect duplicates from SNPs
}
