% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_process.R
\name{get.miss}
\alias{get.miss}
\title{Get missingness of individuals in raw vcf}
\usage{
get.miss(vcf, plot = TRUE)
}
\arguments{
\item{vcf}{data frame of imported vcf file using readVCF}

\item{plot}{logical. Whether to plot the missingness density with a suggested cut-off.}
}
\description{
A function to get missingness of snps data on a per-individual basis similar to --missing-indiv option in vcftools
}
\examples{
vcf.file.path <- paste0(path.package("rCNV"), "/example.raw.vcf.gz")
vcf <- readVCF(vcf.file.path=vcf.file.path)
missing<-get.miss(vcf,plot=TRUE)

}
\author{
Piyal Karunarathne
}
