% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{h.zygosity}
\alias{h.zygosity}
\title{Determine per sample heterozygosity and inbreeding coefficient}
\usage{
h.zygosity(vcf, plot = FALSE, pops = NA, verbose = TRUE)
}
\arguments{
\item{vcf}{an imported vcf file in data.frame or matrix format using "readVCF"}

\item{plot}{logical. Whether to plot a boxplot of inbreeding coefficients for populations. A list of populations must be provided}

\item{pops}{character. A list of population names}

\item{verbose}{logical. Show progress}
}
\value{
Returns a data frame of expected "E(Hom)", observed "O(Hom)" homozygotes with their inbreeding coefficients.
}
\description{
This function will calculate the heterozygosity on a per-sample basis from vcf files (snps), and most importantly inbreeding coefficient which is used to filter out the samples with bad mapping quality. See details.
}
\examples{
vcf.file.path <- paste0(path.package("rCNV"), "/example.raw.vcf.gz")
vcf <- readVCF(vcf.file.path=vcf.file.path)
hzygots<-h.zygosity(vcf,plot=TRUE,pops=c("FR_PA","RU_PA"))

}
\author{
Piyal Karunarathne Pascal Milesi
}
