% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hets}
\alias{hets}
\title{Heterozygosity example data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 10000 rows and 194 columns.
}
\usage{
data(hets)
}
\description{
example SNPs data of Picea abies from Chen et al. 2019
The data contains only a partial snps data set of exome-capture data after filtering
}
\references{
Chen et al. 2019 (add the correct reference)
}
\keyword{datasets}
