% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMM_normalz.R
\name{normz}
\alias{normz}
\title{Calculate normalized depth for total depth}
\usage{
normz(
  df,
  method = c("TMM", "TMMex"),
  logratioTrim = 0.3,
  sumTrim = 0.05,
  Weighting = TRUE,
  Acutoff = -1e+10,
  tot.count = 1e+06
)
}
\arguments{
\item{df}{a data frame or matrix of count values (total counts per snp per sample)}

\item{method}{character. method to be used (see detials). Default="TMM"}

\item{logratioTrim}{numeric. percentage value (0 - 1) of variation to be trimmed in log transformation}

\item{sumTrim}{numeric. amount of trim to use on the combined absolute levels ("A" values) for method="TMM"}

\item{Weighting}{logical, whether to compute (asymptotic binomial precision) weights}

\item{Acutoff}{numeric, cutoff on "A" values to use before trimming}

\item{tot.count}{numeric, a total library size that depth values to be normalized to, default=1000000}
}
\value{
Returns a data frame of normalized depth values
}
\description{
This function outputs the normalized depth values calculated using normalization factor with trimmed mean of M-values of sample libraries. See details.
}
\details{
This function converts an observed depth value table to an effective depth vallue table using TMM normalization. See the original publication for more information.
}
\examples{
print("to be added")

}
\references{
Robinson MD, Oshlack A (2010). A scaling normalization method for differential expression analysis of RNA-seq data. Genome Biology 11, R25
Robinson MD, McCarthy DJ and Smyth GK (2010). edgeR: a Bioconductor package for differential expression analysis of digital gene expression data. Bioinformatics 26, 139-140
}
\author{
Piyal Karunarathne
}
