% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_process.R
\name{readVCF}
\alias{readVCF}
\title{Import VCF file}
\usage{
readVCF(vcf.file.path, verbose = FALSE)
}
\arguments{
\item{vcf.file.path}{path to the vcf file}

\item{verbose}{logical. show progress}
}
\value{
Returns a list with vcf table in a data frame, excluding meta data.
}
\description{
Function to import raw single and multi-sample VCF files.
The function required the R-package \code{data.table} for faster importing.
}
\examples{
\dontrun{vcf.file.path <- paste0(path.package("rCNV"), "/example.raw.vcf.gz")
vcf <- readVCF(vcf.file.path)}

}
\author{
Piyal Karunarathne
}
