% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.als.R
\name{sig.snps}
\alias{sig.snps}
\title{Detect duplicated snps from depth of coverage and number of heterozygotes}
\usage{
sig.snps(dup.info, plot = TRUE)
}
\arguments{
\item{dup.info}{data.frame of duplication info generated by dup.snp.info function}

\item{plot}{logical. whether to plot the duplicated and singleton snps}
}
\value{
a dataframe of snps with their status of duplication
}
\description{
This function will determine if a snp is duplicated or singleton based on the number of heterozygotes present for that specific snps and the median depth of coverage for the allele.
}
\examples{
data(hets)
dup.info <- dup.snp.info(het.table=hets,normalize=FALSE)
duplicated<-sig.snps(dup.info)

}
\author{
Piyal Karunarathne
}
