% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_detect.R
\name{vst}
\alias{vst}
\title{Calculate population wise Vst}
\usage{
vst(AD, pops, id.list = NULL, qGraph = TRUE, ...)
}
\arguments{
\item{AD}{data frame of total allele depth values of (duplicated if id.list is not provided) SNPs}

\item{pops}{character. A vector of population names for each individual. Must be the same length as the number of samples in AD}

\item{id.list}{character. A vector of duplicated SNP IDs. Must match the IDs in the AD data frame}

\item{qGraph}{logical. Plot the network plot based on Vst values (see details)}

\item{...}{additional arguments passed to qgraph}
}
\value{
returns a matrix of pairwise Vst values for populations
}
\description{
This function calculates Vst (......) for populations given a list of duplicated loci
}
\details{
more details on the qgraph
}
\examples{
#to be added

}
\author{
Piyal Karunarathne
}
