% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADnorm}
\alias{ADnorm}
\title{Normalized allele depth example data}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(ADnorm)
}
\description{
Normalized example SNPs data of Chinook Salmon from Larson et al. 2014
The data has been normalized with TMM
}
\references{
Larson, W. A., Seeb, L. W., Everett, M. V., Waples, R. K., Templin, W. D., & Seeb, J. E. (2014). Genotyping by sequencing resolves shallow population structure to inform conservation of Chinook salmon (Oncorhynchus tshawytscha). Evolutionary Applications, 7(3), 355-369
}
\keyword{datasets}
