% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADtable}
\alias{ADtable}
\title{Allele Depth (AD) example data}
\format{
An object of class \code{data.frame} with 5518 rows and 150 columns.
}
\usage{
data(ADtable)
}
\description{
example SNPs data of Picea abies from Chen et al. 2019
The data contains only a partial snps data set of exome-capture data after filtering
}
\references{
McKinney et al. 2017
}
\keyword{datasets}
