% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{cnv}
\alias{cnv}
\title{Find CNVs from deviants}
\usage{
cnv(
  data,
  test = c("z.het", "z.05", "z.all", "chi.het", "chi.05", "chi.all"),
  plot = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame of allele information generated with the function
\code{allele.info}}

\item{test}{vector of characters. Type of test to be used for significance.
See details}

\item{plot}{logical. Plot the detection of duplicates. default \code{TRUE}}

\item{verbose}{logical. show progress}

\item{...}{other arguments to be passed to \code{plot}}
}
\value{
Returns a data frame of SNPs with their detected duplication status
}
\description{
Categorize deviant and non-deviant into "singlets" and "duplicates" based on the statistical approaches specified by the user.
The intersection of all the stats provided will be used in the categorization. If one would like to use the intersection of at least two stats, this can be specified in the \code{n.ints}
}
\details{
SNP deviants are detected with both excess of heterozygosity
according to HWE and deviant SNPs where depth values fall outside of the
normal distribution are detected using the
following methods:
\enumerate{
\item Z-score test \eqn{Z = \frac{\frac{N}{2} -  N_{A}}{\sigma _{x}}}
\item chi-square test (see references for more details on the method)
}

Users can pick among Z-score for heterozygotes (\code{z.het, chi.het}),
all allele combinations (\code{z.all, chi.all}) and the assumption of no
probe bias p=0.5 (\code{z.05, chi.05})
}
\examples{
\dontrun{data(alleleINF)
DD<-cnv(alleleINF)}

}
\author{
Piyal Karunarathne
}
