% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMM_normalz.R
\name{cpm.normal}
\alias{cpm.normal}
\title{Calculate normalized depth for alleles}
\usage{
cpm.normal(
  het.table,
  method = c("TMM", "TMMex", "MedR", "QN"),
  logratioTrim = 0.3,
  sumTrim = 0.05,
  Weighting = TRUE,
  Acutoff = -1e+10,
  verbose = TRUE
)
}
\arguments{
\item{het.table}{a data frame or matrix of coverage (output of hetTgen with "AD")}

\item{method}{character. method to be used (see detials). Default="TMM"}

\item{logratioTrim}{numeric. percentage value (0 - 1) of variation to be trimmed in log transformation}

\item{sumTrim}{numeric. amount of trim to use on the combined absolute levels ("A" values) for method="TMM"}

\item{Weighting}{logical, whether to compute (asymptotic binomial precision) weights}

\item{Acutoff}{numeric, cutoff on "A" values to use before trimming}

\item{verbose}{logical. show progress}
}
\value{
Returns a data frame of normalized depth values similar to the output of hetTgen function
}
\description{
This function outputs the normalized depth values separately for each allele, calculated using normalization factor with trimmed mean of M-values of sample libraries, or median ratios normalization or .... See details.
}
\details{
This function converts an observed depth value table to an effective depth vallue table using TMM normalization (See the original publication for more information). It is different from the function normz only in calculation of the counts per million is for separate alleles instead of the total depth. The "TMMex" method is an extension of the "TMM" method for large data sets containing SNPs exceeding 10000. The method "MedR" is median ratio normalization (see publication); QN - quantile normalization (see publication).
}
\examples{
data(ADtable)
ADnormalized<-cpm.normal(ADtable)


}
\references{
Robinson MD, Oshlack A (2010). A scaling normalization method for differential expression analysis of RNA-seq data. Genome Biology 11, R25
Robinson MD, McCarthy DJ and Smyth GK (2010). edgeR: a Bioconductor package for differential expression analysis of digital gene expression data. Bioinformatics 26, 139-140
}
\author{
Piyal Karunarathne, Qiujie Zhou
}
