% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.als.R
\name{depthVsSample}
\alias{depthVsSample}
\title{Simulate median allele ratios for varying no. of samples and depth coverage}
\usage{
depthVsSample(cov.len = 400, sam.len = 1000, incr = c(1, 1))
}
\arguments{
\item{cov.len}{Max value of depth of coverage to be simulated}

\item{sam.len}{Maximum no. of samples to be simulated}

\item{incr}{A vector of two integers indicating ncrement size for both depth and no. samples ranges}
}
\value{
A matrix of median allele rations where rows are the number of samples and columns are depth of coverage values
}
\description{
This function will simulate the expected median allele ratios under HWE for given ranges of no. of samples and depth coverage values.
This is useful if you need to find the cutoff values of allele ratios for different no. of samples and depth of coverage values in your dataset.
}
\examples{

depthVsSample(cov.len=10,sam.len=50)

}
