% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{dup.plot}
\alias{dup.plot}
\title{Plot classified SNPs into deviants/CNVs and non-deviants/non-CNVs}
\usage{
dup.plot(ds, ...)
}
\arguments{
\item{ds}{a data frame of detected deviants/cnvs (outputs of functions above)}

\item{\dots}{other graphical parameters to be passed to the function
\code{plot}}
}
\value{
Returns no value, only plots proportion of heterozygotes vs allele
median ratio separated by duplication status
}
\description{
The function plots detected deviants/CNVs from functions \code{sig.snps},
\code{cnv} and \code{dupGet} on a median ratio (MedRatio) Vs. proportion of
heterozygote (PropHet) plot.
}
\examples{
\dontrun{data(alleleINF)
DD<-dupGet(alleleINF,plot=FALSE)
dup.plot(DD)}

}
\author{
Piyal Karunarathne
}
