% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{dup.plot}
\alias{dup.plot}
\title{Plot duplicates}
\usage{
dup.plot(ds, ...)
}
\arguments{
\item{ds}{a data frame of detected duplicates}

\item{\dots}{other graphical parameters to be passed to the function
\code{plot}}
}
\value{
Returns no value, only plots proportion of heterozygotes vs allele
median ratio seperated by duplication status
}
\description{
The function plots detected duplicates from functions \code{sig.snps}, and
\code{dupGet}
}
\examples{
\dontrun{data(alleleINF)
DD<-dupGet(alleleINF,plot=FALSE)
dup.plot(DD)}

}
\author{
Piyal Karunarathne
}
