% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{dup.plot}
\alias{dup.plot}
\title{Plot duplicates}
\usage{
dup.plot(ds, ...)
}
\arguments{
\item{ds}{a data frame of detected duplicates}

\item{...}{other graphical parameters to be passed to the function plot}
}
\value{
plots duplicates on proportion of heterozygotes vs allele median ratio
}
\description{
The function plots detected duplicates from functions "sig.snps", and "dup.detect"
}
\examples{
print("to be added with new ones")

}
\author{
Piyal Karunarathne
}
