% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup.snp.info.R
\name{dup.snp.info}
\alias{dup.snp.info}
\title{Generate SNPs duplication information}
\usage{
dup.snp.info(het.table, normalize = FALSE, verbose = TRUE)
}
\arguments{
\item{het.table}{tab-delimited depth value by sample/population (output of hetTgen)}

\item{normalize}{logical. Whether to apply normalization to depth values using trimmed mean of M-values. See ref.}

\item{verbose}{logical. if TRUE, shows the progress}
}
\description{
This function generates duplication information of all the SNPs provided in the data set.
dup.snp.info takes a tab separated depth values Vs. SNPs by sample/population table and calculates numbers and proportions of hetero/homo-zygotes for each snp.
The resulting table is used in the downstream analysis such as CNV/duplication detection.
}
\examples{
print("deprecated")

}
\references{
Robinson MD, Oshlack A (2010). A scaling normalization method for differential expression analysis of RNA-seq data. Genome Biology 11, R25.
}
\author{
Piyal Karunarathne
}
