% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_detect.R
\name{dup.validate}
\alias{dup.validate}
\title{Validate detected duplicates}
\usage{
dup.validate(d.detect, window.size = 100)
}
\arguments{
\item{d.detect}{a data frame of detected SNPs of duplicates/singletons (output of dup.detect)}

\item{window.size}{numerical. a single value of the desired moving window size (default=100 bp)}
}
\value{
a data frame of scaffold names and their average presence in the scaffold.
}
\description{
This function will validate the detected duplicated-SNPs using a moving window approach (see details)
}
\author{
Piyal Karunarathne
}
