% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_detect.R
\name{dup.validate}
\alias{dup.validate}
\title{Validate detected duplicates}
\usage{
dup.validate(d.detect, window.size = 100, scaf.size = 10000)
}
\arguments{
\item{d.detect}{a data frame of detected SNPs of duplicates and singlets
(output of \code{dupGet})}

\item{window.size}{numerical. a single value of the desired moving window
size (default \code{100} bp)}

\item{scaf.size}{numerical. scaffold size to be checked. i.e. the split size of the fragment to be checked with the specified window size.
default=10000}
}
\value{
A data frame of scaffold names and their average presence in the
scaffold.
}
\description{
This function will validate the detected duplicated-SNPs using a moving
window approach (see details)
}
\details{
Chromosome positions correctly ordered according to a reference
sequence is necessary for this function to work properly. Therefore, this
function is still in development for non-mapped reference sequences.
}
\author{
Piyal Karunarathne
}
