% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_process.R
\name{get.miss}
\alias{get.miss}
\title{Get missingness of individuals in raw vcf}
\usage{
get.miss(data, type = c("samples", "snps"), plot = TRUE, verbose = TRUE)
}
\arguments{
\item{data}{a list containing imported vcf file using readVCF or genotype table generated using hetTgen}

\item{type}{character. Get the missing percentages per sample or per SNP."samples" or "snps", default both}

\item{plot}{logical. Whether to plot the missingness density with 95\% quantile}

\item{verbose}{logical. Whether to show progress}
}
\description{
A function to get missingness of snps data on a per-individual basis similar to --missing-indiv option in vcftools
}
\examples{
vcf.file.path <- paste0(path.package("rCNV"), "/example.raw.vcf.gz")
vcf <- readVCF(vcf.file.path=vcf.file.path)
missing<-get.miss(vcf,plot=TRUE)

}
\author{
Piyal Karunarathne
}
