% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_process.R
\name{gt.format}
\alias{gt.format}
\title{Format genotype for BayEnv and BayPass}
\usage{
gt.format(gt, info, snp.subset = FALSE, verbose = FALSE)
}
\arguments{
\item{gt}{multi-vector. an imported data.frame of genotypes or genotype data frame generated by "hetTgen" or path to GT.FORMAT file generated from VCFTools}

\item{info}{a data frame containing sample and population information. It must have "sample" and "population" columns}

\item{snp.subset}{logical. whether to generate a randomly sampled subset of 1/10 snps}

\item{verbose}{logical. If TRUE shows progress}
}
\value{
Returns a list with formated genotype data: $hor - snps in horizontal format (two lines per snp); $ver - vertical format (two column per snp); $hor.chunk - a subset snps of $hor
}
\description{
This function generates necessary genotype count formats for BayEnv and BayPass with a subset of SNPs
}
\examples{
\dontrun{vcf.file.path <- paste0(path.package("rCNV"), "/example.raw.vcf.gz")
vcf <- readVCF(vcf.file.path=vcf.file.path)
het.table<-hetTgen(vcf,"GT")
info<-unique(substr(colnames(het.table)[-c(1:3)],1,8))
GT<-gt.format(het.table,info)}

}
\author{
Piyal Karunarathne
}
