% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_process.R
\name{hetTgen}
\alias{hetTgen}
\title{Generate heterozygote table (allele depth values)}
\usage{
hetTgen(vcf, info.type = c("AD", "GT", "GT-012", "GT-AB"), verbose = TRUE)
}
\arguments{
\item{vcf}{an imported vcf file in data.frame or matrix format using "readVCF"}

\item{info.type}{character. "AD"=allele depth value, "GT"=genotype. Default "AD". See details.}

\item{verbose}{logical. whether to show the progress of the analysis}
}
\description{
hetTgen extracts the read depth and coverage values for each snp for all the individuals from a vcf file generated from readVCF (or GatK VariantsToTable: see details)
}
\details{
If you generate the depth values for allele by sample using GatK VariantsToTable option, use only -F CHROM -F POS -GF AD flags to generate the table. Or keep only the CHROM POS and individual AD columns.
For info.type "GT" option is provided to extract the genotypes of individuals by snp. However to obtain the heterozygosity data necessary for getting the duplication information with the function dup.snp.info, info.type needs to be set to "AD", the default argument.
}
\examples{
vcf.file.path <- paste0(path.package("rCNV"), "/example.raw.vcf.gz")
vcf <- readVCF(vcf.file.path=vcf.file.path)
het.table<-hetTgen(vcf)

}
\author{
Piyal Karunarathne
}
