% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_process.R
\name{maf}
\alias{maf}
\title{Remove MAF allele}
\usage{
maf(h.table, AD = TRUE, verbose = TRUE, parallel = FALSE)
}
\arguments{
\item{h.table}{allele depth table generated from the function \code{hetTgen}}

\item{AD}{logical. If TRUE a allele depth table similar to \code{hetTgen}
output will be returns; If \code{FALSE}, individual AD values per SNP will be
returned in a list.}

\item{verbose}{logical. Show progress}

\item{parallel}{logical. whether to parallelize the process}
}
\value{
A data frame or a list of minimum allele frequency removed allele depth
}
\description{
A function to remove the alleles with minimum allele frequency and keep only
a bi-allelic matrix when loci are multi-allelic
}
\examples{
\dontrun{mf<-maf(ADtable)}

}
\author{
Piyal Karunarathne
}
