% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_process.R
\name{maf}
\alias{maf}
\title{Remove MAF allele}
\usage{
maf(h.table, AD = TRUE, verbose = TRUE)
}
\arguments{
\item{h.table}{allele depth table generated from the function \link[rCNV]{hetTgen}}

\item{AD}{logical. If TRUE a allele depth table similar to \link[rCNV]{hetTgen} output will be returns; If FALSE, individual AD values per SNP will be returned in a list.}

\item{verbose}{logical. Show progress}
}
\value{
a data frame or a list
}
\description{
A function to remove the alleles with minimum allele frequency and keep only a biallelic matrix
}
\examples{
mf<-maf(ADtable)

}
\author{
Piyal Karunarathne
}
