% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_process.R
\name{readVCF}
\alias{readVCF}
\title{Import VCF file}
\usage{
readVCF(vcf.file.path)
}
\arguments{
\item{vcf.file.path}{path to the vcf file}
}
\value{
Returns a dataframe with scaffold/chromosome positions and depth values for individual samples
}
\description{
Function to import raw single/multi-sample VCF files generated from GatK or VCFtools.
The function required the R-package "data.table" for faster importing.
}
\examples{
vcf.file.path <- paste0(path.package("rCNV"), "/example.raw.vcf")
vcf <- readVCF(vcf.file.path=vcf.file.path)

}
\author{
Piyal Karunarathne
}
