% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{sig.hets}
\alias{sig.hets}
\title{Identify significantly different heterozygotes from SNPs data}
\usage{
sig.hets(
  a.info,
  method = c("fisher", "chi.sq"),
  plot = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{a.info}{allele info table generated from filtered vcfs using the function 'allele.info()'}

\item{method}{character. Method for testing significance. Fisher exact test ("fisher") or Chi squre test ("chi.sq")}

\item{plot}{logical. Whether to plot the identified duplicated snps with the expected values}

\item{verbose}{logical, if TRUE, the progress is shown}

\item{...}{other arguments passed to plot}
}
\value{
A matrix of expected heterozygote proportions from the observed data with p-value indicating significantly deviating snps, thus duplicates.
If enabled, the function also plots the recognized duplicates in red and expected (singletons) in black on a prop. of homozygote alt Vs. prop. of heterozygote plot
p2 observed homozygote reference
het expected
}
\description{
This function will recognize the SNPs that are significantly different from the expected under HWE and plot potential duplicated snps
}
\examples{
\dontrun{data(alleleINF)
AI <- alleleINF
duplicates<-sig.hets(AI,plot=TRUE)}

}
\author{
Piyal Karunarathne, Pascal Milesi
}
