% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.als.R
\name{sig.hets}
\alias{sig.hets}
\title{Identify significantly different heterozygotes from SNPs data}
\usage{
sig.hets(dup.info, plot = TRUE)
}
\arguments{
\item{dup.info}{duplication info table generated from filtered vcfs using the function dup.snp.info}

\item{plot}{logical. Wheather to plot the identified duplicated snps with the expected values}
}
\description{
This function will recognize the SNPs that are significantly different from the expected under HWE and plot potential duplicated snips
}
\examples{
data(dup.info)
duplicates<-sig.hets(dup.info,plot=TRUE)

}
