% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.als.R
\name{sig.hets}
\alias{sig.hets}
\title{Identify significantly different heterozygotes from SNPs data}
\usage{
sig.hets(d.info, plot = TRUE)
}
\arguments{
\item{d.info}{duplication info table generated from filtered vcfs using the function dup.snp.info}

\item{plot}{logical. Whether to plot the identified duplicated snps with the expected values}
}
\value{
A matrix of expected heterozygote proportions from the observed data with p-value indicating significantly deviating snps, thus duplicates.
If enabled, the function also plots the recognized duplicates in red and expected (singletons) in black on a prop. of homozygote alt Vs. prop. of heterozygote plot
p2 observed homozygote reference
het expected
}
\description{
This function will recognize the SNPs that are significantly different from the expected under HWE and plot potential duplicated snps
}
\examples{
data(hets)
d.info <- dup.snp.info(het.table=hets,normalize=FALSE)
duplicates<-sig.hets(d.info,plot=TRUE)

}
\author{
Piyal Karunarathne, Pascal Milesi
}
