% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.als.R
\name{sig.snps}
\alias{sig.snps}
\title{Detect duplicated snps from depth of coverage and number of heterozygotes}
\usage{
sig.snps(d.info, stringency = c("95", "99", "max"), plot = TRUE)
}
\arguments{
\item{d.info}{a data.frame of duplication info generated by dup.snp.info function}

\item{stringency}{character. Confidence interval to be used in filtering duplicates}

\item{plot}{logical. whether to plot the duplicated and singleton snps}
}
\value{
a dataframe of snps with their status of duplication. If plot=TRUE, plot depicting duplicated and non-duplicated (singleton) alleles
}
\description{
This function will determine if a snp is duplicated or singleton based on the number of heterozygotes present for that specific snps and the median depth of coverage for the allele.
}
\examples{
data(hets)
d.info <- dup.snp.info(het.table=hets,normalize=FALSE)
duplicated<-sig.snps(d.info,stringency="max")

}
\author{
Piyal Karunarathne
}
